+++ IP12-Prototype/src/main/ch/IP12/prototype/model/Obstacle.java
@@ -20,7 +20,7 @@ public class Obstacle extends Moveable{
     }
 
     @Override
-    public void update(double deltaTime, float strength){
+    public void update(double deltaTime, double strength){
         //forces the direction of the obstacle to change over time
         if (waveUp){
             direction--;
